Added configparser, removed contextlib
Added configparser and an 'actual' config.ini by request on the temp folder.

Removed contextlib because suppress was skipping the moves instead of actually suppressing the errors.

Seems to take a little while longer to move files, not sure if its a FS or memory issue. Still not impossibly slow, but it is noticeable.